(refactor) Refactor server to route db requests
